Create PepWitch1.5.onlineandoffline
A few updates here:
1) Fixed the string parsing so that the names of the protein IDs load correctly
2) Added a major offline mode that allows for easier back-end processing and utility checking

This will be the last major update for version 1. Right now, the code is so messy that I'm considering re-writing the whole thing so that it's easier to read. The biggest outstanding issue is that of NSAF calculation in relation to the original replicate states. Right now, PepWitch calculates NSAF based off filtered values instead of the original replicate values. SAFs work best when they serve as a representation of the original replicate state.